+++ super_editor/lib/src/default_editor/document_hardware_keyboard/document_keyboard_actions.dart
@@ -200,7 +200,7 @@ ExecutionInstruction doNothingWhenThereIsNoSelection({
   }
 }
 
-ExecutionInstruction doNothinOnMac({
+ExecutionInstruction sendKeyEventToMacOs({
   required SuperEditorContext editContext,
   required RawKeyEvent keyEvent,
 }) {
